feat: nav graph 파일 분리, network module 수정, appversion utl 추가,
