+++ app/src/main/java/com/example/zzanz_android/MainActivity.kt
@@ -9,21 +9,8 @@ import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.core.view.WindowCompat
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navigation
-import com.example.zzanz_android.common.NavRoutes
-import com.example.zzanz_android.common.SettingNavRoutes
-import com.example.zzanz_android.common.SpendingNavRoutes
+import com.example.zzanz_android.common.navigation.NavHost
 import com.example.zzanz_android.common.ui.theme.ZzanZ_AndroidTheme
-import com.example.zzanz_android.presentation.view.Home
-import com.example.zzanz_android.presentation.view.Setting
-import com.example.zzanz_android.presentation.view.Splash
-import com.example.zzanz_android.presentation.view.spending.AddSpendingScreen
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -39,60 +26,6 @@ class MainActivity : ComponentActivity() {
                     NavHost()
                 }
             }
-        }
-    }
-}
-
-fun NavGraphBuilder.settingGraph(navController: NavHostController) {
-    navigation(
-        startDestination = SettingNavRoutes.Budget.route,
-        route = NavRoutes.Setting.route
-    ) {
-        composable(SettingNavRoutes.Budget.route) {
-            Setting(navController = navController, route = SettingNavRoutes.Budget.route)
-        }
-        composable(SettingNavRoutes.BudgetByCategory.route) {
-            Setting(navController = navController, route = SettingNavRoutes.BudgetByCategory.route)
-        }
-        composable(SettingNavRoutes.BudgetCategory.route) {
-            Setting(navController = navController, route = SettingNavRoutes.BudgetCategory.route)
-        }
-        composable(SettingNavRoutes.AlarmSetting.route) {
-            Setting(navController = navController, route = SettingNavRoutes.AlarmSetting.route)
-        }
-    }
-}
-
-// TODO : gowoon - Navigation Graph 파일 생성
-fun NavGraphBuilder.spendingGraph(navController: NavHostController){
-    navigation(
-        startDestination = SpendingNavRoutes.AddSpending.route,
-        route = NavRoutes.Spending.route
-    ){
-        composable(SpendingNavRoutes.AddSpending.route){
-            AddSpendingScreen(navController)
-        }
-    }
-}
-
-@Composable
-fun NavHost(
-    modifier: Modifier = Modifier,
-    navController: NavHostController = rememberNavController(),
-    startDestination: String = NavRoutes.Splash.route
-) {
-    NavHost(
-        modifier = modifier.fillMaxSize(),
-        navController = navController,
-        startDestination = startDestination
-    ) {
-        settingGraph(navController = navController)
-        spendingGraph(navController = navController)
-        composable(NavRoutes.Home.route) {
-            Home(navController)
-        }
-        composable(NavRoutes.Splash.route) {
-            Splash(navController)
         }
     }
 }
