+++ app/src/main/java/com/example/zzanz_android/ZzanZApp.kt
@@ -2,6 +2,12 @@ package com.example.zzanz_android
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
 
 @HiltAndroidApp
-class ZzanZApp : Application()
+class ZzanZApp : Application(){
+    override fun onCreate() {
+        super.onCreate()
+        Timber.plant(Timber.DebugTree())
+    }
+}
