+++ app/src/main/java/com/example/zzanz_android/data/di/NetworkModule.kt
@@ -2,11 +2,8 @@ package com.example.zzanz_android.data.di
 
 import android.content.Context
 import com.example.zzanz_android.BuildConfig
-import com.example.zzanz_android.data.remote.api.ChallengeService
-import com.example.zzanz_android.data.remote.api.ChallengeServiceImpl
-import com.example.zzanz_android.data.repository.ChallengeRepositoryImpl
+import com.example.zzanz_android.data.util.AppVersionUtil
 import com.example.zzanz_android.data.util.AuthorizationManager
-import com.example.zzanz_android.domain.repository.ChallengeRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -16,26 +13,34 @@ import io.ktor.client.HttpClient
 import io.ktor.client.engine.cio.CIO
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.plugins.logging.DEFAULT
 import io.ktor.client.plugins.logging.LogLevel
 import io.ktor.client.plugins.logging.Logger
 import io.ktor.client.plugins.logging.Logging
 import io.ktor.client.request.header
+import io.ktor.http.URLProtocol
+import io.ktor.http.path
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
+import timber.log.Timber
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object NetworkModule {
-    private const val apiVersion = "/v1"
+    private const val apiVersion = "v1/"
     @Singleton
     @Provides
     fun provideKtorClient(@ApplicationContext applicationContext: Context): HttpClient {
         return HttpClient(CIO){
             defaultRequest {
-                url(BuildConfig.BASE_URL + apiVersion)
+                url {
+                    protocol = URLProtocol.HTTP
+                    host = BuildConfig.BASE_URL
+                    port = 8080
+                    path(apiVersion)
+                }
                 header("Authorization", AuthorizationManager.getDeviceId(applicationContext))
+                header("App-Version", AppVersionUtil.getVersionHeader())
             }
             install(ContentNegotiation) {
                 json(Json {
@@ -45,7 +50,11 @@ object NetworkModule {
                 })
             }
             install(Logging){
-                logger = Logger.DEFAULT
+                logger = object : Logger {
+                    override fun log(message: String) {
+                        Timber.d("Ktor Logger", message)
+                    }
+                }
                 level = LogLevel.ALL
             }
         }
