--- app/src/main/java/com/example/zzanz_android/presentation/view/HomeScreen.kt
@@ -1,55 +1,0 @@
-package com.example.zzanz_android.presentation.view
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
-import com.example.zzanz_android.common.NavRoutes
-
-@Composable
-fun Home(navController: NavHostController) {
-    // NavHostController TestCode
-    Surface(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(20.dp)
-    ) {
-        Column(
-            modifier = Modifier.fillMaxSize()
-        ) {
-            Text(
-                text = "Home Screen", textAlign = TextAlign.Center, fontSize = 20.sp
-            )
-            Spacer(modifier = Modifier.height(20.dp))
-            Row(
-                modifier = Modifier.wrapContentWidth(),
-                horizontalArrangement = Arrangement.spacedBy(16.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                TestNavButton(route = NavRoutes.Splash.route, navController = navController)
-                TestNavButton(route = NavRoutes.Home.route, navController = navController)
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun HomePreview() {
-    Home(navController = rememberNavController())
-}
